// frontend/src/config/generated-chains.ts
export interface ChainConfig {
    id: number;
    name: string;
    key: string;
    logoURI?: string;
    nativeToken: {
        symbol: string;
        decimals: number;
        name: string;
        logoURI?: string;
        address: string;
    };
    metamask: {
        chainId: string;
        chainName: string;
        nativeCurrency: {
            name: string;
            symbol: string;
            decimals: number;
        };
        rpcUrls: string[];
        blockExplorerUrls: string[];
    };
}

export const evmChains: ChainConfig[] = [
    {
// ...
        "metamask": {
            "chainId": "0x1",
            "blockExplorerUrls": [
                "https://etherscan.io/"
            ],
            "chainName": "Ethereum Mainnet",
            "nativeCurrency": {
                "name": "ETH",
                "symbol": "ETH",
                "decimals": 18
            },
            "rpcUrls": [
                "https://ethereum-rpc.publicnode.com",
                "https://eth.drpc.org"
            ]
        }
    },
// ...
];